Fix: handle empty list in select_from_values_for_batch_range
The current implementation of select_from_values_for_batch_range
constructs an empty VALUES () clause in case it's passed an empty
list for its 'values' argument, which is usually a syntax error.

This commit makes it so we instead return a dummy query that's
guaranteed to not return any rows (using a WHERE FALSE filter).

Additionally, it fixes a few bugs in the unit test module related
to test cases that contain empty input / output rows.