feat: improve logging
previously, it was a common flow to have users rerun plans with --debug
in order for the core team to get more comprehensive logs. this pr trims
down logs and also logs all relevant sql queries as info to the file so
that by default we log essential information for debugging.