feat: improve slack formatting
Slack allows more structured notifications than a console or an email.
This changes the base class so it passes down more information from the
notification to concrete implementations. For slack, it will convert
this information into a more structure message.

For other text-based targets, the anther base class is now available.