fix: reenable depends on self check
add a new depends on self property that doesn't include manual
depends_on, only those that are actually referenced in the query.

a manual depends_on without an actual self reference means that a query
can still operate with a ctas so we don't need to raise an error.

additionally depends_on_past was incorrect because
incremental_by_unique_key also depends_on_past but can be ctas'd because
it doesn't self reference