fix: Fix test runner not displaying dots for successes
Since the introduction of the custom method `addSuccess` for
`ModelTextTestResult` in d8ff3fd4c8946ab756609fc38c6d43552aea103e, the
test runner stopped displaying dots for successful test cases.

This is due to the logging being part of the base implementation itself:

    # unittest.TextTestResult::addSuccess
    self.stream.write('.')
    self.stream.flush()

This patch adds a call the superclass implementation of `addSucess` to
achieve the original behavior of reporting successes.