Refactor!: normalize environment names according to the target engine's dialect
We previously treated environment names as case-insensitive, because we converted
them into lowercase regardless of the target engine.

This was problematic for engines like Snowflake, because essentially forced users
to use quotes in queries that referenced tables within these environments. So, for
example instead of FOO_DEV, we'd create FOO_dev.

This PR addresses this issue by normalizing the environment names according to the
rules of the targeted engine's dialect. So, in the above example we'll now produce
FOO_DEV and users will be able to query FOO_DEV.<table> without using quotes.