fix: issue with stringify pydantic models with sets
pydantic models with nested sets of other pydantic models cannot be
called with .dict(). we call dict() on equality checks in order to
recursively check equality. this is trigger on stringify because we
check that each field is not equal to the default value. this commit
bypasses that check so we don't trigger the dict() calls.

related github issue https://github.com/pydantic/pydantic/issues/8016