Fix star macro when used with snowflake
The star macro mismatches the case of the table alias
when used with snowflake dialect.
The table alias will be uppercase while
the alias used with columns remains lowercase.
Since the identifiers are all quoted this
causes a problem for snowflake.

Solution is to apply normalization function to the table_identifier.